+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 * Support for defining the course name using the ``cert_name_long`` field (in Studio's Advanced Settings).
 * Support for specifying individual fonts for PDF text elements.
 * Support for \n in learning context names in PDF certificates.
+* Options for uppercase name and issue date in PDF certificates.
 
 Modified
 ========